// src/app/enum/operetor.enum.ts
export enum FilterOperetorEnum {
    And = 'and',
    Or = 'or',
}

export enum OperetorEnum {
    Equal = '=',
    NotEqual = '!=',
    MoreThan = '>',
    LessThan = '<',
    MoreThanOrEqual = '>=',
    LessThanOrEqual = '<=',
    ILike = 'like',
    // NotLike = 'not like',
    In = 'in',
    // NotIn = 'not in',
    IsNull = 'is null',
    IsNotNull = 'is not null',
    Between = 'between',
}